requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
python-multipart>=0.0.12

# LLM Providers
openai>=1.50.0
google-generativeai>=0.8.0
httpx>=0.27.0

# Vector store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# Document parsing
pymupdf4llm>=0.0.10
python-docx>=1.1.0

# Database
aiosqlite>=0.20.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.5.0

# CLI
typer[all]>=0.12.0
rich>=13.9.0

# Logging
loguru>=0.7.0
